drm: eswin: reject framebuffers with misaligned pitch
Currently the pitch of dumb buffers seems to be not respected by the IMG
blob, which leads to framebuffers with misaligned pitch get created; in
addition setting such a broken pitch to the STRIDE register seems to
crash the DC and stops it from scanning out.

Filter out such framebuffers when attaching them to KMS planes.

Signed-off-by: Icenowy Zheng <uwu@icenowy.me>